Create core_mini_axi_sim
- Create a new target -- core_mini_axi_sim -- which is a testbench to
  execute binaries in the CoreMiniAxi variant of Kelvin.
- Utilizes libsystemctlm-soc to bridge CoreMiniAxi's AXI ports to the
  testbench via TLM.
- Has a few fake peripherals that can be accessed via the master port.
- HDL fixes to execute correctly with the AXI signals generated by
  libsystemctlm-soc.

+++ tests/renode/rv_core/kelvin_hello_world.c
@@ -19,7 +19,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-volatile uint32_t* uart0 = (uint32_t*)0x54000000L;
+volatile uint8_t* uart0 = (uint8_t*)0x54000000L;
 void putc(char ch) {
   *uart0 = ch;
 }
